Organize the code
Move the maincontent to the orderpage to make it organize.
